State: split (initial) reducer into a separate module
The code that constructs the (initial) reducer is big enough to deserve its
own module, so this patch creates a `state/reducer` one.

This move will become even more useful once we start dynamically loading reducers.
Then the thing exported from `state/reducer` will be just the initial value. It's
already apparent in this patch that the `serialize` logic will need to change a bit.
